+++ packages/extension/preview/evaluate-in-webview.ts
@@ -1,7 +1,6 @@
 import * as fs from 'fs';
 
-import { mdxTranspileAsync } from '../transpiler/mdx/mdx';
-import { transformAsync as babelTransformAsync } from '../transpiler/babel';
+import { transformEntry } from '../module-fetcher/transform';
 
 const precinct = require("precinct");
 
@@ -17,17 +16,9 @@ export default async function evaluateInWebview(preview: Preview, text: string, 
   try {
     performance.mark('preview/start');
 
-    const { languageId, uri } = preview.doc;
-    let code: string;
-    if (languageId === 'markdown' || languageId === 'mdx' || uri.scheme === 'untitled') {
-      code = await mdxTranspileAsync(text, true, preview);
-    } else {
-      code = text;
-    }
-
+    const code = await transformEntry(text, fsPath, preview);
     const entryFilePath = fs.realpathSync(fsPath);
     const entryFileDependencies = precinct(code);
-    code = (await babelTransformAsync(code)).code;
 
     console.log(code);
     console.log(entryFilePath);