+++ packages/extension/transpiler/babel.ts
@@ -19,9 +19,3 @@ const babelOptions = {
 export const transformAsync = code => {
   return babel.transformAsync(code, babelOptions);
 };
-
-// TODO: provide a preference to use sucrase
-// import { transform } from "sucrase";
-// export const transformAsync = code => {
-//   return transform(code, { transforms: ["jsx", "typescript", "imports"] });
-// };