+++ packages/extension/transpiler/sucrase.ts
@@ -1,0 +1,5 @@
+import { transform as sucraseTransform } from "sucrase";
+
+export const transform = (code: string) => {
+  return sucraseTransform(code, { transforms: ["jsx", "typescript", "imports"] });
+};