+++ packages/webview-app/src/rpc-webview-handle.ts
@@ -1,9 +1,7 @@
-import { evaluate } from './evaluate';
+import { evaluate, invalidate } from './evaluate';
 import {
-  currentBuildError,
   startReportingRuntimeErrorsIfNotStarted,
   handleBuildError,
-  dismissBuildError
 } from './lib/errors';
 
 class RPCWebviewHandle {
@@ -11,15 +9,16 @@ class RPCWebviewHandle {
 
   updatePreview(code: string, entryFilePath: string, entryFileDependencies: Array<string>) {
     startReportingRuntimeErrorsIfNotStarted();
-    if (currentBuildError) {
-      dismissBuildError();
-    }
     evaluate(code, entryFilePath, entryFileDependencies);
   }
 
   showPreviewError(error: Error) {
     startReportingRuntimeErrorsIfNotStarted();
     handleBuildError(error);
+  }
+
+  async invalidate(fsPath: string) {
+    await invalidate(fsPath);
   }
 }